Moved out BuildingData
