+++ src/Engine/Data/CMakeLists.txt
@@ -6,7 +6,9 @@ set(ENGINE_DATA_HEADERS
         AutonoteEnums.h
         AutonoteData.h
         AwardData.h
-        AwardEnums.h)
+        AwardEnums.h
+        BuildingData.h
+        BuildingEnums.h)
 
 add_library(engine_data INTERFACE ${ENGINE_DATA_SOURCES} ${ENGINE_DATA_HEADERS})
 target_link_libraries(engine_data INTERFACE library_serialization utility)
