+++ src/Engine/Objects/Character.cpp
@@ -29,6 +29,7 @@
 #include "Engine/SpellFxRenderer.h"
 #include "Engine/AttackList.h"
 #include "Engine/Tables/AwardTable.h"
+#include "Engine/Tables/BuildingTable.h"
 #include "Engine/Tables/ItemTable.h"
 #include "Engine/Tables/CharacterFrameTable.h"
 #include "Engine/Tables/StorylineTextTable.h"
