+++ src/Engine/Objects/Character.h
@@ -5,6 +5,7 @@
 #include <string>
 #include <utility>
 
+#include "Engine/Data/BuildingEnums.h"
 #include "Engine/Objects/NPCEnums.h"
 #include "Engine/Objects/ActorEnums.h"
 #include "Engine/Objects/CombinedSkillValue.h"
@@ -14,7 +15,6 @@
 #include "Engine/Objects/MonsterEnums.h"
 #include "Engine/Spells/SpellEnums.h"
 #include "Engine/Spells/SpellBuff.h"
-#include "Engine/Tables/BuildingTable.h"
 #include "Engine/Events/EventEnums.h"
 #include "Engine/Pid.h"
